# OKComputer-3D/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 3D Print Business Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [TODO, IN_PROGRESS, REVIEW, DONE, CANCELLED]
        - name: priority
          in: query
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, URGENT]
        - name: projectId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, PLANNING, IN_PROGRESS, ON_HOLD, COMPLETED, CANCELLED]
        - name: clientId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quotes
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, SENT, ACCEPTED, REJECTED, EXPIRED]
        - name: clientId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a new quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "201":
          description: Created quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a quote by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: Update a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteBody"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [TODO, IN_PROGRESS, REVIEW, DONE, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
        dueDate:
          type: ["string", "null"]
          format: date-time
        projectId:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - tags
        - createdAt
        - updatedAt
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [TODO, IN_PROGRESS, REVIEW, DONE, CANCELLED]
          default: TODO
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
          default: MEDIUM
        dueDate:
          type: ["string", "null"]
          format: date-time
        projectId:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
      required:
        - title
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [TODO, IN_PROGRESS, REVIEW, DONE, CANCELLED, null]
        priority:
          type: ["string", "null"]
          enum: [LOW, MEDIUM, HIGH, URGENT, null]
        dueDate:
          type: ["string", "null"]
          format: date-time
        projectId:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [DRAFT, PLANNING, IN_PROGRESS, ON_HOLD, COMPLETED, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        progress:
          type: integer
        startDate:
          type: ["string", "null"]
          format: date-time
        deadline:
          type: ["string", "null"]
          format: date-time
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - priority
        - progress
        - tags
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [DRAFT, PLANNING, IN_PROGRESS, ON_HOLD, COMPLETED, CANCELLED]
          default: DRAFT
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
          default: MEDIUM
        clientId:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
          format: date-time
        deadline:
          type: ["string", "null"]
          format: date-time
        tags:
          type: array
          items:
            type: string
      required:
        - name
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [DRAFT, PLANNING, IN_PROGRESS, ON_HOLD, COMPLETED, CANCELLED, null]
        priority:
          type: ["string", "null"]
          enum: [LOW, MEDIUM, HIGH, URGENT, null]
        clientId:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        progress:
          type: ["integer", "null"]
        startDate:
          type: ["string", "null"]
          format: date-time
        deadline:
          type: ["string", "null"]
          format: date-time
        tags:
          type: array
          items:
            type: string
...
    Quote:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [DRAFT, SENT, ACCEPTED, REJECTED, EXPIRED]
        clientId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        projectId:
          type: ["string", "null"]
        totalAmount:
          type: number
        validUntil:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - totalAmount
        - items
        - createdAt
        - updatedAt
    QuoteItem:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        total:
          type: number
      required:
        - id
        - description
        - quantity
        - unitPrice
        - total
    CreateQuoteBody:
      type: object
      properties:
        title:
          type: string
        clientId:
          type: ["string", "null"]
        projectId:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuoteItemBody"
      required:
        - title
    CreateQuoteItemBody:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
      required:
        - description
        - quantity
        - unitPrice
    UpdateQuoteBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [DRAFT, SENT, ACCEPTED, REJECTED, EXPIRED, null]
        clientId:
          type: ["string", "null"]
        projectId:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuoteItemBody"
...