# OKComputer-3D/lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Print Business Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyPriority } from "./updateProjectBodyPriority";
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  status?: UpdateProjectBodyStatus;
  /** @nullable */
  priority?: UpdateProjectBodyPriority;
  /** @nullable */
  clientId?: string | null;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  startDate?: Date | null;
  /** @nullable */
  deadline?: Date | null;
  tags?: string[];
}