# OKComputer-3D/lib/api-zod/src/generated/types/updateQuoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Print Business Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateQuoteItemBody } from "./createQuoteItemBody";
import type { UpdateQuoteBodyStatus } from "./updateQuoteBodyStatus";

export interface UpdateQuoteBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  status?: UpdateQuoteBodyStatus;
  /** @nullable */
  clientId?: string | null;
  /** @nullable */
  projectId?: string | null;
  /** @nullable */
  validUntil?: Date | null;
  /** @nullable */
  notes?: string | null;
  items?: CreateQuoteItemBody[];
}